// prisma/generated/type-graphql/resolvers/crud/Endofshift/args/UpsertOneEndofshiftArgs.d.ts
import { EndofshiftCreateInput } from "../../../inputs/EndofshiftCreateInput";
import { EndofshiftUpdateInput } from "../../../inputs/EndofshiftUpdateInput";
import { EndofshiftWhereUniqueInput } from "../../../inputs/EndofshiftWhereUniqueInput";
export declare class UpsertOneEndofshiftArgs {
    where: EndofshiftWhereUniqueInput;
    create: EndofshiftCreateInput;
    update: EndofshiftUpdateInput;
}